(fix) use double field

+++ src/gws_biota/compound/compound.py
@@ -21,7 +21,7 @@
 # The licensor cannot revoke these freedoms as long as you follow the license terms.
 # https://creativecommons.org/licenses/by/4.0/.
 
-from peewee import CharField, DecimalField, IntegerField, ForeignKeyField
+from peewee import CharField, FloatField, DoubleField, IntegerField, ForeignKeyField
 from peewee import Model as PeeweeModel
 
 from gws_core.model.typing_register_decorator import typing_registrator
@@ -48,9 +48,9 @@ class Compound(Base):
     kegg_id = CharField(null=True, index=True)
     metacyc_id = CharField(null=True, index=True)
     formula = CharField(null=True, index=True)
-    charge = DecimalField(null=True, index=True)
-    mass = DecimalField(null=True, index=True)
-    monoisotopic_mass = DecimalField(null=True, index=True)
+    charge = FloatField(null=True, index=True)
+    mass = DoubleField(null=True, index=True)
+    monoisotopic_mass = DoubleField(null=True, index=True)
     inchi = CharField(null=True, index=True)
     inchikey = CharField(null=True, index=True)
     smiles = CharField(null=True, index=True)
