+++ src/gws_biota/reaction/reaction_position.py
@@ -3,7 +3,7 @@
 # The use and distribution of this software is prohibited without the prior consent of Gencovery SAS.
 # About us: https://gencovery.com
 
-from peewee import CharField, DecimalField
+from peewee import CharField, FloatField
 from gws_core.model.typing_register_decorator import typing_registrator
 from gws_core import BadRequestException, JSONField
 from ..base.base import Base
@@ -12,9 +12,9 @@ from .reaction_position_data import REACTION_POSITION_DATA
 @typing_registrator(unique_name="ReactionPosition", object_type="MODEL", hide=True)
 class ReactionPosition(Base):
     rhea_id = CharField(null=True, index=True)
-    x = DecimalField(null=True, index=True)
-    y = DecimalField(null=True, index=True)
-    z = DecimalField(null=True, index=True)
+    x = FloatField(null=True, index=True)
+    y = FloatField(null=True, index=True)
+    z = FloatField(null=True, index=True)
     points = JSONField(null=True, index=False)
 
     def __init__(self, *args, **kwargs):
